+++ types/src/index.ts
@@ -1,0 +1,4 @@
+export * from './schema'
+
+import * as jsonSchema from './generated/schema.json'
+export { jsonSchema }